+++ controllers/ResidentsController.php
@@ -63,30 +63,35 @@ class ResidentsController extends Controller {
      */
     public function actionCreate() {
         $model = new Residents();
-        $model->scenario = 'create';
+        //$model->scenario = 'create';
 
         if ($model->load(Yii::$app->request->post())) {
 
-            //INSTACIO EL ARCHIVO CARGADO
-            if (!$model->file = \yii\web\UploadedFile::getInstance($model, 'file')) {
-                Yii::$app->session->setFlash('error', "El archivo no pudo "
-                        . "ser cargado. Inténtelo de nuevo.");
-                return $this->redirect(['index']);
+            if ($model->file) {
+                //INSTACIO EL ARCHIVO CARGADO
+                if (!$model->file = \yii\web\UploadedFile::getInstance($model, 'file')) {
+                    Yii::$app->session->setFlash('error', "El archivo no pudo "
+                            . "ser cargado. Inténtelo de nuevo.");
+                    return $this->redirect(['index']);
+                }
+
+
+                //RUTA DE ALMACENAJE LOCAL Y NOMBRE
+                $ruta = 'archivos/' . date('YmdHis') . '-'
+                        . strtolower(trim(str_replace($this->especial, $this->wespecial, $model->file->baseName))) . '.'
+                        . strtolower(trim($model->file->extension));
+                if (!@$model->file->saveAs($ruta, false)) {
+                    Yii::$app->session->setFlash('error', "El archivo no pudo "
+                            . "ser guardado. Inténtelo de nuevo.");
+                    return $this->redirect(['index']);
+                }
+
+                //GUARDO LOS DATOS
+                $model->photo = $ruta;
+            } else {
+                $model->photo = "";
             }
 
-
-            //RUTA DE ALMACENAJE LOCAL Y NOMBRE
-            $ruta = 'archivos/' . date('YmdHis') . '-'
-                    . strtolower(trim(str_replace($this->especial, $this->wespecial, $model->file->baseName))) . '.'
-                    . strtolower(trim($model->file->extension));
-            if (!@$model->file->saveAs($ruta, false)) {
-                Yii::$app->session->setFlash('error', "El archivo no pudo "
-                        . "ser guardado. Inténtelo de nuevo.");
-                return $this->redirect(['index']);
-            }
-
-            //GUARDO LOS DATOS
-            $model->photo = $ruta;
             if (!$model->save()) {
                 unlink($model->photo);
                 Yii::$app->session->setFlash('error', "El archivo no pudo "
@@ -125,11 +130,13 @@ class ResidentsController extends Controller {
                     return $this->redirect(['index']);
                 }
                 //GUARDO LOS DATOS
-                unlink($model->photo);
+                if(file_exists($model->photo)){
+                    unlink($model->photo);
+                }                
                 $model->photo = $ruta;
             }
 
-            if (!$model->save()) {                
+            if (!$model->save()) {
                 Yii::$app->session->setFlash('error', "El archivo no pudo "
                         . "ser cargado. Inténtelo de nuevo.");
                 return $this->redirect(['index']);
